# Backend/swagger.yml
openapi: 3.0.3
info:
  title: Concert Manager API
  description: |
    API REST pour la gestion de concerts, groupes musicaux et salles.
    
    Cette API permet de :
    - Gérer les groupes musicaux (CRUD, import Excel)
    - Gérer les salles de concert (CRUD, import Excel)
    - Programmer et gérer les concerts (CRUD)
    - Consulter les statistiques du dashboard
    - Consulter les activités récentes
    

    **Format** : JSON

  version: 1.0.0
  contact:
    name: API Support
    email: support@concertmanager.com
  license:
    name: MIT
servers:
  - url: http://localhost:8005/api
    description: Serveur de développement local
  - url: https://api.concertmanager.com/api
    description: Serveur de production
...
paths:
...
  /groups:
    get:
      tags:
        - Groups
      summary: Liste des groupes avec filtres et pagination
      description: |
        Récupère la liste des groupes musicaux avec possibilité de filtrer par :
        - name : Nom du groupe (recherche partielle)
        - city : Ville d'origine
        - musicalStyle : Style musical
        - startYear : Année de début
        
        La pagination est gérée via les paramètres `page` et `limit`.
      operationId: listGroups
      parameters:
        - name: name
          in: query
          description: Filtrer par nom (recherche partielle)
          required: false
          schema:
            type: string
        - name: city
          in: query
          description: Filtrer par ville
          required: false
          schema:
            type: string
        - name: musicalStyle
          in: query
          description: Filtrer par style musical
          required: false
          schema:
            type: string
        - name: startYear
          in: query
          description: Filtrer par année de début
          required: false
          schema:
            type: integer
            minimum: 1900
            maximum: 2100
        - name: page
          in: query
          description: "Numéro de page (défaut: 1)"
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: "Nombre d'éléments par page (défaut: 10, max: 100)"
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Liste des groupes récupérée avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupListResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    post:
      tags:
        - Groups
      summary: Créer un nouveau groupe
      description: Crée un nouveau groupe musical avec les informations fournies
      operationId: createGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupCreateRequest'
            example:
              name: "The Beatles"
              origin: "Liverpool, UK"
              city: "Liverpool"
              startYear: 1960
              endYear: 1970
              founders: "John Lennon, Paul McCartney, George Harrison, Ringo Starr"
              membersCount: 4
              musicalStyle: "Rock"
              presentation: "Groupe de rock britannique légendaire"
      responses:
        '201':
          description: Groupe créé avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /groups/{id}:
    get:
      tags:
        - Groups
      summary: Détails d'un groupe
      description: Récupère les détails complets d'un groupe par son ID
      operationId: getGroup
      parameters:
        - name: id
          in: path
          required: true
          description: ID du groupe
          schema:
            type: integer
      responses:
        '200':
          description: Détails du groupe récupérés avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    put:
      tags:
        - Groups
      summary: Modifier un groupe
      description: Met à jour les informations d'un groupe existant
      operationId: updateGroup
      parameters:
        - name: id
          in: path
          required: true
          description: ID du groupe
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupUpdateRequest'
      responses:
        '200':
          description: Groupe modifié avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    delete:
      tags:
        - Groups
      summary: Supprimer un groupe
      description: Supprime un groupe et tous ses concerts associés
      operationId: deleteGroup
      parameters:
        - name: id
          in: path
          required: true
          description: ID du groupe
          schema:
            type: integer
      responses:
        '200':
          description: Groupe supprimé avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /groups/import:
    post:
      tags:
        - Groups
      summary: Importer des groupes depuis un fichier Excel
      description: |
        Importe plusieurs groupes depuis un fichier Excel (.xlsx, .xls).
        
        Format attendu du fichier Excel :
        - Colonne A : Nom
        - Colonne B : Origine
        - Colonne C : Ville
        - Colonne D : Année début
        - Colonne E : Année séparation
        - Colonne F : Fondateurs
        - Colonne G : Nombre de membres
        - Colonne H : Style musical
        - Colonne I : Présentation
      operationId: importGroups
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Fichier Excel (.xlsx ou .xls)
      responses:
        '200':
          description: Import réussi
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "3 groupes importés avec succès"
                  data:
                    type: object
                    properties:
                      imported:
                        type: integer
                        example: 3
                      errors:
                        type: array
                        items:
                          type: object
                          properties:
                            row:
                              type: integer
                            message:
                              type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    Group:
      type: object
      properties:
        id:
          type: integer
          description: ID unique du groupe
          example: 1
        name:
          type: string
          description: Nom du groupe
          example: "The Beatles"
        origin:
          type: string
          description: Origine du groupe
          example: "Liverpool, UK"
        city:
          type: string
          nullable: true
          description: Ville d'origine
          example: "Liverpool"
        startYear:
          type: integer
          nullable: true
          description: Année de début
          minimum: 1900
          maximum: 2100
          example: 1960
        endYear:
          type: integer
          nullable: true
          description: Année de séparation
          minimum: 1900
          maximum: 2100
          example: 1970
        founders:
          type: string
          nullable: true
          description: Fondateurs du groupe
          example: "John Lennon, Paul McCartney, George Harrison, Ringo Starr"
        membersCount:
          type: integer
          description: Nombre de membres
          minimum: 1
          example: 4
        musicalStyle:
          type: string
          description: Style musical
          example: "Rock"
        presentation:
          type: string
          nullable: true
          description: Présentation du groupe
          example: "Groupe de rock britannique légendaire"
        createdAt:
          type: string
          format: date-time
          description: Date de création
          example: "2026-01-15T10:30:00+00:00"
    GroupCreateRequest:
      type: object
      required:
        - name
        - origin
        - membersCount
        - musicalStyle
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: "The Beatles"
        origin:
          type: string
          minLength: 1
          maxLength: 255
          example: "Liverpool, UK"
        city:
          type: string
          maxLength: 255
          nullable: true
          example: "Liverpool"
        startYear:
          type: integer
          minimum: 1900
          maximum: 2100
          nullable: true
          example: 1960
        endYear:
          type: integer
          minimum: 1900
          maximum: 2100
          nullable: true
          example: 1970
        founders:
          type: string
          nullable: true
          example: "John Lennon, Paul McCartney, George Harrison, Ringo Starr"
        membersCount:
          type: integer
          minimum: 1
          example: 4
        musicalStyle:
          type: string
          minLength: 1
          maxLength: 255
          example: "Rock"
        presentation:
          type: string
          nullable: true
          example: "Groupe de rock britannique légendaire"
    GroupUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/GroupCreateRequest'
        - type: object
          properties:
            id:
              type: integer
              description: ID du groupe à modifier
    GroupResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Group'
    GroupListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Group'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # ===========================================
    # SALLE SCHEMAS
    # ===========================================
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          description: Page actuelle
          example: 1
        limit:
          type: integer
          description: Nombre d'éléments par page
          example: 10
        total:
          type: integer
          description: Nombre total d'éléments
          example: 100
        totalPages:
          type: integer
          description: Nombre total de pages
          example: 10
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Opération réussie"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Message d'erreur détaillé"
  responses:
    BadRequest:
      description: Requête invalide (400)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Les données fournies sont invalides"
    NotFound:
      description: Ressource introuvable (404)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Ressource introuvable"
    InternalServerError:
      description: Erreur serveur (500)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Une erreur interne s'est produite"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Authentification par token JWT.
        Actuellement en mode mock pour le développement.
        Format: `Bearer {token}`
security:
  - BearerAuth: []